Limited the return of searches and who can delete whom

+++ src/main/java/com/quizdeck/controllers/UserRequestController.java
@@ -1,15 +1,21 @@
 package com.quizdeck.controllers;
 
+import com.quizdeck.exceptions.ForbiddenAccessException;
 import com.quizdeck.exceptions.InvalidJsonException;
 import com.quizdeck.model.database.User;
+import com.quizdeck.model.responses.UserSearchOutput;
 import com.quizdeck.repositories.UserRepository;
+import io.jsonwebtoken.Claims;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import javax.validation.Valid;
+import java.util.ArrayList;
 
 /**
  * Created by Cade on 3/11/2016.
@@ -17,6 +23,8 @@ import javax.validation.Valid;
 @RestController
 @RequestMapping("/rest/secure/user")
 public class UserRequestController {
+
+    private Logger log = LoggerFactory.getLogger(UserRequestController.class);
 
     @Autowired
     UserRepository userRepository;
@@ -32,15 +40,30 @@ public class UserRequestController {
     }
 
     @RequestMapping(value="/deleteUser", method=RequestMethod.DELETE)
-    public ResponseEntity<String> deleteUser(@PathVariable String userId){
+    public ResponseEntity<String> deleteUser(@ModelAttribute("claims") Claims claims, @PathVariable String userId)throws ForbiddenAccessException{
+        User temp = userRepository.findByUserName(claims.get("user").toString());
+        if(temp.getId() != userId) {
+            throw new ForbiddenAccessException();
+        }
         userRepository.delete(userId);
         return new ResponseEntity<String>(HttpStatus.OK);
     }
 
-
     @RequestMapping(value="/findUser/{userName}", method = RequestMethod.GET)
-    public User findUser(@PathVariable String userName){
-        return userRepository.findByUserName(userName);
+    public ArrayList<UserSearchOutput> findUser(@PathVariable String userName){
+
+        ArrayList<User> users = userRepository.findByUserNameLike(userName);
+        ArrayList<UserSearchOutput> out = new ArrayList<>();
+        if(users == null || users.size() == 0) {
+            log.info("query returned nothing");
+        }
+        else{
+            for (User user : users) {
+                out.add(new UserSearchOutput(user.getUserName()));
+            }
+        }
+
+        return out;
     }
 
 }
